+++ src/features/source/browse/screens/SourceMangas.tsx
@@ -423,10 +423,10 @@ export function SourceMangas() {
         <>
             <AppbarSearch />
             <SourceGridLayout />
-            <CustomTooltip title={t`Open in WebView`} disabled={!source?.baseUrl}>
+            <CustomTooltip title={t`Open in WebView`} disabled={!source?.homeUrl}>
                 <IconButton
-                    disabled={!source?.baseUrl}
-                    href={source?.baseUrl ? requestManager.getWebviewUrl(source?.baseUrl) : ''}
+                    disabled={!source?.homeUrl}
+                    href={source?.homeUrl ? requestManager.getWebviewUrl(source?.homeUrl) : ''}
                     rel="noreferrer"
                     target="_blank"
                     color="inherit"