+++ src/features/source/services/Sources.ts
@@ -14,7 +14,7 @@ import type {
     SourceLanguageInfo,
     SourceNsfwInfo,
     SourceMetaInfo,
-    SourceRepoInfo,
+    SourceStoreInfo,
 } from '@/features/source/Source.types.ts';
 import {
     DefaultLanguage,
@@ -34,6 +34,7 @@ import type { SourceBaseFieldsFragment } from '@/lib/graphql/generated/graphql.t
 import { requestManager } from '@/lib/requests/RequestManager.ts';
 import type { DocumentNode, Unmasked } from '@apollo/client';
 import { SOURCE_BASE_FIELDS } from '@/lib/graphql/source/SourceFragments.ts';
+import { isNsfw as isNsfwFnc } from '@/features/extension/Extensions.utils.ts';
 
 export class Sources {
     static readonly LOCAL_SOURCE_ID = '0';
@@ -134,7 +135,7 @@ export class Sources {
         const normalizedLanguages = toComparableLanguages(toUniqueLanguageCodes(languages ?? []));
 
         const filters: [Condition: any, CheckKeepLocalSource: boolean, Filter: (source: Source) => boolean][] = [
-            [isNsfw, true, (source: Source) => source.isNsfw === isNsfw],
+            [isNsfw, true, (source: Source) => isNsfwFnc(source.contentWarning) === isNsfw],
             [
                 languages,
                 true,
@@ -157,14 +158,14 @@ export class Sources {
         }, sources);
     }
 
-    static areFromMultipleRepos<Source extends SourceIdInfo & SourceRepoInfo>(sources: Source[]): boolean {
-        const repo = sources.find((source) => !!source.extension.repo)?.extension.repo;
+    static areFromMultipleStores<Source extends SourceIdInfo & SourceStoreInfo>(sources: Source[]): boolean {
+        const store = sources.find((source) => !!source.extension.storeIndexUrl)?.extension.storeIndexUrl;
 
-        if (!repo || !sources.length) {
+        if (!store || !sources.length) {
             return false;
         }
 
-        return sources.some((source) => source.extension.repo !== repo && !Sources.isLocalSource(source));
+        return sources.some((source) => source.extension.storeIndexUrl !== store && !Sources.isLocalSource(source));
     }
 
     static getLastUsedSource<Source extends SourceIdInfo & SourceMetaInfo>(